More work for 1.3
flagsBitfield was removed from the Header class, and was replaced with an EntryInfo for it.

Volatile patches, One time patches, and flagBitfield flags are now handled via the flagIsEnabled function, in which helper functions are used to access them easier.

Renamed WALLETS_PATCHED to INCREASE_WALLETS to better match the generator.

Replaced datetime.utcnow in gcipack.py since its deprecated.