Remove Randomizer pointer from Seed constructor params
gRandomizer is now initialized in the Randomizer constructor, so a pointer to it is no longer needed to be passed into the Seed constructor.

Also adjusted how the message about the seed successfully being applied is written in the console, in which the seed is on its own line, to avoid it potentially being cut off due to its length.